WIP : Refactor feature.upnp.content.processor package

+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/persistence/dialect/AnsiAlbumDao.java
@@ -46,7 +46,7 @@ public class AnsiAlbumDao implements DialectAlbumDao {
     }
 
     @Override
-    public List<Album> getAlbumsByGenre(int offset, int count, List<String> genres,
+    public List<Album> getAlbumsByGenre(long offset, long count, List<String> genres,
             List<MusicFolder> folders) {
         if (genres.isEmpty() || folders.isEmpty()) {
             return Collections.emptyList();
